DBZ-5283 Move ExtractChangedRecordState tests to own class
